Added net::contentLength() to query "content-length" field from HTTP header.
Adds net::contentLength(). This makes a short HTTP request to read the
"content-length" field from the HTTP header.

This will be used to determine the size of a file before downloading
it to the fetcher cache, in order to ensure there is enough space
ahead of time.

Review: https://reviews.apache.org/r/30606